Added FireBall class for player attack and fixed scratch attack
Added FireBall as an entity of MapObject. Stored in an arraylist until
it hits something and "dies".   fixed the animation of scratch attack.
noticed that the gliding is bugged, once you glide once you are always
gliding, possibly related to using Shift key as input, or it's in the
keyevent logic, will fix that later.